+++ src/avdecc/controllerModel.hpp
@@ -40,6 +40,8 @@ public:
 		GptpDomain,
 		InterfaceIndex,
 		AssociationId,
+		MediaClockMasterId,
+		MediaClockMasterName,
 
 		Count
 	};