+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Display entities' media clock master id and name in the entity list (Contributed by d&b)
+- Support for media clock master determination (Contributed by d&b)
+
 ### Changed
 - Icon when an entity is acquired by Hive (changed color from orange to green)
 
@@ -71,7 +75,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [1.0.3] - 2018-07-17
 ### Added
-- Support for Memory Object descriptors (Contributed by Florian Harmuth)
+- Support for Memory Object descriptors (Contributed by d&b)
 - Settings menu
 
 ### Fixed